Upgrade fatfs to R0.10a

+++ libs/fat_sd/sdio_diskio.c
@@ -74,7 +74,7 @@ DRESULT disk_read (
   BYTE drv, /* Physical drive number (0) */
   BYTE *buff, /* Pointer to the data buffer to store read data */
   DWORD sector, /* Start sector number (LBA) */
-  BYTE count /* Sector count (1..255) */
+  UINT count /* Sector count */
   )
 {
   uint16_t Transfer_Length;
@@ -98,7 +98,7 @@ DRESULT disk_write (
   BYTE drv, /* Physical drive number (0) */
   const BYTE *buff, /* Pointer to the data to be written */
   DWORD sector, /* Start sector number (LBA) */
-  BYTE count /* Sector count (1..255) */
+  UINT count /* Sector count */
   )
 {
   uint16_t Transfer_Length;
